Fixes "Invalid calling object" on IE11
The global `toString` method in IE does not allow it to be called on
objects other than `window`. It will throw an "Invalid calling object"
error when this happens.

This PR ensures that the `Object.prototype.toString` is used instead
which allows any value to be converted to a string.